move choice header to new component
