+++ src/components/choices/ChoiceDetail.js
@@ -9,21 +9,16 @@ import { OptionContext } from "../options/OptionProvider"
 import { Option } from "../options/Option"
 import Table from "react-bootstrap/Table"
 import { Rating } from "../ratings/Rating"
-import Button from "react-bootstrap/esm/Button"
-import { ChoiceDelete } from "./ChoiceDelete"
-import { ModalContext } from "../modals/ModalProvider"
+import { ChoiceHeader } from "./ChoiceHeader"
 
 export const ChoiceDetail = (props) => {
-  const { choices, getChoices, editChoice, } = useContext(ChoiceContext)
+  const { choices, getChoices } = useContext(ChoiceContext)
   const { factors, getFactors } = useContext(FactorContext)
   const { options, getOptions } = useContext(OptionContext)
-  const { handleShow, setModalContent } = useContext(ModalContext)
   const [ choice, setChoice ] = useState({})
   const [ choiceFactors, setChoiceFactors ] = useState([])
   const [ choiceOptions, setChoiceOptions ] = useState([])
-  const [ isChoiceNameChanging, setIsChoiceNameChanging ] = useState(false)
 
-  const choiceRef = useRef(null)
 
   useEffect( () => {
     getChoices()
@@ -40,59 +35,19 @@ export const ChoiceDetail = (props) => {
       setChoiceOptions(choiceOptions)
   }, [choices, factors, options])
 
-  const changeName = () => {
-    const newChoice = choice
-    newChoice.name = choiceRef.current.value 
-    editChoice(newChoice)
-    setIsChoiceNameChanging(false)
-  }
+  
 
   return <> 
   <section className="choice">
-    <Button 
-        onClick={evt => {
-          props.history.push("/")
-        }}>
-        Back
-    </Button>
-    <div className="choice__name">
-      { 
-      (isChoiceNameChanging)
-      ? <>
-      <input type="text" defaultValue={choice.name} ref={choiceRef}></input>
-      <Button 
-            onClick={evt => {
-                changeName()
-            }}>
-            Save
-        </Button>
-      </>
-      : <>
-      <div onClick={evt => {
-          setIsChoiceNameChanging(true)
-        }}
-        className="btn">
-          {choice.name}
-      </div>
-      </>
-      }
-    </div>
-    <Button 
-        onClick={evt => {
-          setModalContent(<ChoiceDelete choice={choice} {...props} />)
-          handleShow()
-        }}>
-        Delete
-    </Button>
     
-    
-    
-    
+    <ChoiceHeader {...props} choice={choice} />
     
     <Table striped bordered >
       <thead>
         <tr>
-          <th><WeightList {...props} /></th>
+          <th>
+            <WeightList {...props} />
+          </th>
           {
             choiceFactors.map(cf => {
               return (
